silence compiler warnings

+++ src/EXTRA-FIX/fix_ave_moments.cpp
@@ -196,7 +196,7 @@ FixAveMoments::FixAveMoments(LAMMPS *lmp, int narg, char **arg) :
   // this fix produces a global vector and array
 
   vector_flag = 1;
-  size_vector = nvalues*moments.size();
+  size_vector = nvalues * moments.size();
   array_flag = 1;
   size_array_rows = size_vector;
   size_array_cols = nhistory;
@@ -239,7 +239,7 @@ FixAveMoments::FixAveMoments(LAMMPS *lmp, int narg, char **arg) :
     else if (extvalue != extarray)
       error->all(FLERR, Error::NOLASTLINE, "Fix ave/moments cannot set output array "
                  "intensive/extensive from these inputs");
-    for (int j=0; j<moments.size(); j++)
+    for (int j=0; j < (int)moments.size(); j++)
       extlist[i + j] = extvalue;
     i += moments.size();
   }
@@ -608,8 +608,8 @@ void FixAveMoments::update_results()
     const double G2 = k4 / square(k2);
 
     // map to result array, starting at value interleave offset
-    double* rfirst = &result[i*moments.size()];
-    for (int j = 0; j < moments.size(); j++) {
+    double* rfirst = &result[i * moments.size()];
+    for (int j = 0; j < (int)moments.size(); j++) {
       switch(moments[j]) {
         case MEAN:
           rfirst[j] = mean;
